+++ src/main/java/book/shop/spring/boot/intro/SpringBootIntroApplication.java
@@ -1,0 +1,12 @@
+package book.shop.spring.boot.intro;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBootIntroApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringBootIntroApplication.class, args);
+    }
+}
