+++ src/main/java/book/shop/spring/boot/intro/dto/CreateBookRequestDto.java
@@ -1,0 +1,14 @@
+package book.shop.spring.boot.intro.dto;
+
+import java.math.BigDecimal;
+import lombok.Data;
+
+@Data
+public class CreateBookRequestDto {
+    private String title;
+    private String author;
+    private String isbn;
+    private BigDecimal price;
+    private String description;
+    private String coverImage;
+}
