+++ src/main/java/book/shop/spring/boot/intro/exception/DataProcessingException.java
@@ -1,0 +1,11 @@
+package book.shop.spring.boot.intro.exception;
+
+public class DataProcessingException extends RuntimeException {
+    public DataProcessingException(String message) {
+        super(message);
+    }
+
+    public DataProcessingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
