+++ src/main/java/book/shop/spring/boot/intro/exception/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package book.shop.spring.boot.intro.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
