+++ src/main/java/book/shop/spring/boot/intro/mapper/BookMapper.java
@@ -1,0 +1,14 @@
+package book.shop.spring.boot.intro.mapper;
+
+import book.shop.spring.boot.intro.config.MapperConfig;
+import book.shop.spring.boot.intro.dto.BookDto;
+import book.shop.spring.boot.intro.dto.CreateBookRequestDto;
+import book.shop.spring.boot.intro.model.Book;
+import org.mapstruct.Mapper;
+
+@Mapper(config = MapperConfig.class)
+public interface BookMapper {
+    BookDto toDto(Book book);
+
+    Book toModel(CreateBookRequestDto bookDto);
+}
