+++ src/main/java/book/shop/spring/boot/intro/repository/BookRepository.java
@@ -1,0 +1,13 @@
+package book.shop.spring.boot.intro.repository;
+
+import book.shop.spring.boot.intro.model.Book;
+import java.util.List;
+import java.util.Optional;
+
+public interface BookRepository {
+    Book save(Book book);
+
+    Optional<Book> findById(Long id);
+
+    List<Book> findAll();
+}
