+++ src/main/java/book/shop/spring/boot/intro/service/BookService.java
@@ -1,0 +1,14 @@
+package book.shop.spring.boot.intro.service;
+
+import book.shop.spring.boot.intro.dto.BookDto;
+import book.shop.spring.boot.intro.dto.CreateBookRequestDto;
+import java.util.List;
+
+public interface BookService {
+
+    BookDto save(CreateBookRequestDto requestDto);
+
+    List<BookDto> findAll();
+
+    BookDto findById(Long id);
+}
