--- src/main/java/mate/academy/spring/boot/intro/config/MapperConfig.java
@@ -1,13 +1,0 @@
-package mate.academy.spring.boot.intro.config;
-
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.NullValueCheckStrategy;
-
-@org.mapstruct.MapperConfig(
-        componentModel = "spring",
-        injectionStrategy = InjectionStrategy.CONSTRUCTOR,
-        nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS,
-        implementationPackage = "<PACKAGE_NAME>.impl"
-)
-public class MapperConfig {
-}
