--- src/main/java/mate/academy/spring/boot/intro/dto/BookDto.java
@@ -1,14 +1,0 @@
-package mate.academy.spring.boot.intro.dto;
-
-import java.math.BigDecimal;
-import lombok.Data;
-
-@Data
-public class BookDto {
-    private Long id;
-    private String title;
-    private String author;
-    private BigDecimal price;
-    private String description;
-    private String coverImage;
-}
