--- src/main/java/mate/academy/spring/boot/intro/dto/CreateBookRequestDto.java
@@ -1,11 +1,0 @@
-package mate.academy.spring.boot.intro.dto;
-
-import java.math.BigDecimal;
-import lombok.Data;
-
-@Data
-public class CreateBookRequestDto {
-    private String title;
-    private String author;
-    private BigDecimal price;
-}
