--- src/main/java/mate/academy/spring/boot/intro/exception/DataProcessingException.java
@@ -1,11 +1,0 @@
-package mate.academy.spring.boot.intro.exception;
-
-public class DataProcessingException extends RuntimeException {
-    public DataProcessingException(String message) {
-        super(message);
-    }
-
-    public DataProcessingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
