--- src/main/java/mate/academy/spring/boot/intro/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package mate.academy.spring.boot.intro.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
