--- src/main/java/mate/academy/spring/boot/intro/model/Book.java
@@ -1,39 +1,0 @@
-package mate.academy.spring.boot.intro.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.math.BigDecimal;
-import lombok.Data;
-import lombok.ToString;
-
-@Data
-@Entity
-@Table(name = "books")
-public class Book {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 255)
-    private String title;
-
-    @Column(nullable = false, length = 255)
-    private String author;
-
-    @ToString.Include
-    @Column(name = "isbn",nullable = true, unique = true)
-    private String isbn;
-
-    @Column(nullable = true, precision = 10, scale = 2)
-    private BigDecimal price;
-
-    @Column(length = 1000)
-    private String description;
-
-    @Column(length = 255)
-    private String coverImage;
-}
