--- src/main/java/mate/academy/spring/boot/intro/repository/BookRepositoryImpl.java
@@ -1,74 +1,0 @@
-package mate.academy.spring.boot.intro.repository;
-
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.EntityTransaction;
-
-import java.util.List;
-import java.util.Optional;
-
-import lombok.RequiredArgsConstructor;
-import mate.academy.spring.boot.intro.dto.BookDto;
-import mate.academy.spring.boot.intro.exception.DataProcessingException;
-import mate.academy.spring.boot.intro.model.Book;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class BookRepositoryImpl implements BookRepository {
-    private final EntityManagerFactory entityManagerFactory;
-
-    @Override
-    public Book save(Book book) {
-        EntityTransaction transaction = null;
-        EntityManager entityManager = entityManagerFactory.createEntityManager();
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            entityManager.persist(book);
-            transaction.commit();
-            return book;
-        } catch (Exception e) {
-            if (transaction != null && transaction.isActive()) {
-                transaction.rollback();
-            }
-            throw new DataProcessingException("Can't insert book into DB: " + book, e);
-        } finally {
-            if (entityManager.isOpen()) {
-                entityManager.close();
-            }
-        }
-    }
-
-    @Override
-    public List<Book> findAll() {
-        try (EntityManager entityManager = entityManagerFactory.createEntityManager()) {
-            return entityManager.createQuery("FROM Book", Book.class).getResultList();
-
-        } catch (Exception e) {
-            throw new DataProcessingException("Can't get all books from FB", e);
-        }
-    }
-
-    @Override
-    public List<Book> findByName(String title) {
-        String lowerCaseName = title.toLowerCase();
-        try(EntityManager entityManager = entityManagerFactory.createEntityManager()) {
-           return entityManager
-                   .createQuery("SELECT e FROM Book e WHERE lower(e.title) LIKE :title", Book.class)
-                   .setParameter("title","%" + lowerCaseName + "%")
-                   .getResultList();
-        }
-    }
-
-    @Override
-    public Optional<Book> findById(Long id) {
-        try (EntityManager entityManager = entityManagerFactory.createEntityManager()) {
-            Book book = entityManager.find(Book.class, id);
-            return book != null ? Optional.of(book) : Optional.empty();
-
-        } catch (Exception e) {
-            throw new DataProcessingException("Can't get book by Id from FB", e);
-        }
-    }
-}
