+++ src/test/java/book/shop/spring/boot/intro/SpringBootIntroApplicationTests.java
@@ -1,0 +1,13 @@
+package book.shop.spring.boot.intro;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringBootIntroApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
