--- src/test/java/mate/academy/spring/boot/intro/SpringBootIntroApplicationTests.java
@@ -1,13 +1,0 @@
-package mate.academy.spring.boot.intro;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringBootIntroApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
